Move CatalogStore to SPI
