+++ core/trino-main/src/main/java/io/trino/connector/CatalogFactory.java
@@ -14,9 +14,11 @@
 package io.trino.connector;
 
 import com.google.errorprone.annotations.ThreadSafe;
+import io.trino.spi.catalog.CatalogProperties;
 import io.trino.spi.connector.CatalogHandle;
 import io.trino.spi.connector.Connector;
 import io.trino.spi.connector.ConnectorFactory;
+import io.trino.spi.connector.ConnectorName;
 
 @ThreadSafe
 public interface CatalogFactory
