+++ core/trino-main/src/main/java/io/trino/connector/ConnectorServicesProvider.java
@@ -14,6 +14,7 @@
 package io.trino.connector;
 
 import io.trino.Session;
+import io.trino.spi.catalog.CatalogProperties;
 import io.trino.spi.connector.CatalogHandle;
 
 import java.util.List;
