+++ core/trino-main/src/main/java/io/trino/execution/CreateCatalogTask.java
@@ -16,13 +16,13 @@ package io.trino.execution;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.inject.Inject;
 import io.trino.Session;
-import io.trino.connector.CatalogName;
-import io.trino.connector.ConnectorName;
 import io.trino.execution.warnings.WarningCollector;
 import io.trino.metadata.CatalogManager;
 import io.trino.metadata.PropertyUtil;
 import io.trino.security.AccessControl;
 import io.trino.spi.TrinoException;
+import io.trino.spi.catalog.CatalogName;
+import io.trino.spi.connector.ConnectorName;
 import io.trino.sql.PlannerContext;
 import io.trino.sql.tree.CreateCatalog;
 import io.trino.sql.tree.Expression;
