+++ core/trino-main/src/main/java/io/trino/execution/DropCatalogTask.java
@@ -15,11 +15,11 @@ package io.trino.execution;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.inject.Inject;
-import io.trino.connector.CatalogName;
 import io.trino.execution.warnings.WarningCollector;
 import io.trino.metadata.CatalogManager;
 import io.trino.security.AccessControl;
 import io.trino.spi.TrinoException;
+import io.trino.spi.catalog.CatalogName;
 import io.trino.sql.tree.DropCatalog;
 import io.trino.sql.tree.Expression;
