+++ core/trino-main/src/main/java/io/trino/metadata/CatalogManager.java
@@ -14,10 +14,10 @@
 package io.trino.metadata;
 
 import com.google.common.collect.ImmutableSet;
-import io.trino.connector.CatalogName;
-import io.trino.connector.CatalogProperties;
-import io.trino.connector.ConnectorName;
+import io.trino.spi.catalog.CatalogName;
+import io.trino.spi.catalog.CatalogProperties;
 import io.trino.spi.connector.CatalogHandle;
+import io.trino.spi.connector.ConnectorName;
 
 import java.util.Map;
 import java.util.Optional;
