+++ core/trino-spi/src/main/java/io/trino/spi/catalog/CatalogProperties.java
@@ -11,9 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.connector;
+package io.trino.spi.catalog;
 
 import io.trino.spi.connector.CatalogHandle;
+import io.trino.spi.connector.ConnectorName;
 
 import java.util.Map;
